refactor: move api order types to `trade::order::api`
