+++ mobile/native/src/api_model/mod.rs
@@ -1,2 +1,1 @@
-pub mod order;
 pub mod position;
